Broken tests, but close to making it work. Need to figure out the Redis connection stuff
Added a supervisor for the periodic task, have to test that it works by setting a shorter interval.

Using cursor-based pagination already. I need to figure out how to make sure that the web template has the ID of the last record in the view so that it can use that for the cursor.