Added a basic migration mix task.
It uses a schema_migrations table that stores the timestamp of the created migration.

It applies the migration in a transaction alongside adding a new schema migrations row.

Migrations are written in pure SQL and stored in the db/ directory.

Next up, I need a migration file generator and figure out how to roll back the migration, though that might not be necessary.